Should be able to pull compute the production chains
I believe that this code is enough to pull a complete quantified list of required buildings for a production chain. However, the quantification will not be correct as the required quantities aren't taken into account. Ie, if a buildings needs 24 items but the supplier only produces 12, the computation will consider that 1 supplier is enough, even tho 2 are needed.